combine escape and unescape of placeholders
