+++ packages/fr-config-pull/src/scripts/amRealmConfig.js
@@ -1,4 +1,4 @@
-const utils = require("../helpers/utils.js");
+const utils = require("../../../fr-config-common/src/utils.js");
 const fs = require("fs");
 const { restGet } = require("../../../fr-config-common/src/restClient.js");
 const { saveJsonToFile } = utils;
