+++ packages/fr-config-pull/src/scripts/variables.js
@@ -4,7 +4,7 @@ const {
   saveJsonToFile,
   esvToEnv,
   escapePlaceholders,
-} = require("../helpers/utils.js");
+} = require("../../../fr-config-common/src/utils.js");
 
 const EXPORT_SUBDIR = "esvs/variables";
